Fix decimal

+++ weather_station/measurements_devices/rain_fall.py
@@ -1,5 +1,4 @@
 from gpiozero import Button
-from decimal import Decimal
 
 rain_sensor = Button(6)
 BUCKET_SIZE = 0.2794
@@ -18,7 +17,7 @@ def reset_rainfall():
 
 def get_data():
     global count
-    return Decimal(str(round(count * BUCKET_SIZE, 2)))
+    return str(round(count * BUCKET_SIZE, 2))
 
 
 rain_sensor.when_pressed = bucket_tipped
